Refactored file structure and file names for StateGraph and files.
